Move CounterExampleConstraint and CounterExampleProgram to geometry::optimization::internal.
